feat: implemented lowering of methods on hir
