+++ crates/hir/src/helpers/mod.rs
@@ -1,4 +1,3 @@
 mod declarations;
 mod expressions;
-mod names;
 mod types;
