+++ crates/hir/src/names.rs
@@ -1,4 +1,4 @@
-use crate::{Result, SlynxHir, SymbolPointer, TypeId, VariableId, module_loader::FileId};
+use crate::{HIRError, Result, SlynxHir, SymbolPointer, TypeId, VariableId, module_loader::FileId};
 
 use common::Span;
 //file specific to implement things related to name resolution
@@ -28,5 +28,22 @@ impl SlynxHir {
         self.types_module.insert_variable(out, ty);
         self.symbols_resolver.create_variable(out, symbol);
         Ok(out)
+    }
+    ///Tries to retrieve a variable with the provided `name` on the current active scope
+    pub fn get_variable(
+        &self,
+        fileid: FileId,
+        symbol: SymbolPointer,
+        span: &Span,
+    ) -> Result<VariableId> {
+        if let Some(variable) = self.get_file(fileid).scopes.get_name(&symbol) {
+            Ok(variable)
+        } else {
+            Err(HIRError::name_unrecognized(symbol, *span))
+        }
+    }
+    ///Retrieves the pointer(simply a symbol) of the provided `name`.
+    pub fn get_symbol(&self, name: &str) -> Option<SymbolPointer> {
+        self.symbols_resolver.retrieve(name)
     }
 }
